% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_metrics}
\alias{model_metrics}
\title{Calculate regression performance metrics}
\usage{
model_metrics(observed, predicted)
}
\arguments{
\item{observed}{Numeric observed values.}

\item{predicted}{Numeric predicted values.}
}
\value{
A one-row data frame.
}
\description{
Calculate regression performance metrics
}
\examples{
model_metrics(observed = c(2, 3, 4), predicted = c(2.1, 2.9, 4.2))
}
