% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_internal_control_benefit}
\alias{plot_internal_control_benefit}
\title{Plot the benefit of internal-control adjustment}
\usage{
plot_internal_control_benefit(sensitivity_data)
}
\arguments{
\item{sensitivity_data}{Data returned by \code{\link[=run_sensitivity_grid]{run_sensitivity_grid()}}.}
}
\value{
A \code{ggplot} heatmap. Negative values favour the adjusted model.
}
\description{
Plot the benefit of internal-control adjustment
}
\examples{
\dontrun{
grid <- run_sensitivity_grid(
  "shared_recovery_sd", c(0.1, 0.4),
  "ic_specific_recovery_sd", c(0.05, 0.3), repetitions = 2L
)
plot_internal_control_benefit(grid)
}
}
