% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the ViralQuantSim Shiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
No return value; launches a Shiny application.
}
\description{
Run the ViralQuantSim Shiny application
}
\examples{
\dontrun{
run_app()
}
}
